# src/main/resources/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "My Demo Spring Boot Application taught by Lukman",
    "version": "1.0.0",
    "title": "Muyi Spring Boot App",
    "contact": {
      "name": "Muyi",
      "email": "mb@gmail.com"
    },
    "license": {
      "name": "Apache License Version 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0\""
    }
  },
  "host": "localhost:8222",
  "basePath": "/",
  "tags": [
    {
      "name": "employee-controller",
      "description": "Employee Controller"
    }
  ],
  "paths": {
    "/v1/employees": {
      "get": {
        "tags": [
          "employee-controller"
        ],
        "summary": "getAllEmployees",
        "operationId": "getAllEmployeesUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "type": "object"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/v1/employees/{emplId}": {
      "get": {
        "tags": [
          "employee-controller"
        ],
        "summary": "getEmployeeById",
        "operationId": "getEmployeeByIdUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "emplId",
            "in": "path",
            "description": "emplId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Employee"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    }
  },
  "definitions": {
    "Employee": {
      "type": "object",
      "properties": {
        "age": {
          "type": "integer",
          "format": "int32"
        },
        "employeeId": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      },
      "title": "Employee"
    }
  }
}